Extracted core of bulk cloning into a library
This is the common code that any clone tool is likely to need. Specifically, this library is usable by the sumdbclone tool that I've got pending and significantly reduces a lot of duplication.
